refactor(#14): 라우팅 구조 리팩토링
