+++ src/Layout/Background/Background.style.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const Container = styled.div``;
