+++ src/Layout/BoardLayout/BoardLayout.jsx
@@ -4,22 +4,16 @@ import {
   MainContentWrapper,
   SidebarWrapper,
   Content,
-  ContentWrapper,
-  BoardTitle,
-  SearchSortWrapper,
-  SearchBox,
-  SortOption,
   RightSidebarWrapper,
   PopularCardWrapper,
-  PostCardWrapper,
 } from "./BoardLayout.style";
 import Sidebar from "../Sidebar/Sidebar";
 import Navbar from "../Navbar/Navbar";
 import PopularHashCard from "../../components/Card/PopularCard/PopularHashCard/PopularHashCard";
 import PopularPostCard from "../../components/Card/PopularCard/PopularPostCard/PopularPostCard";
-import { sortPostsByCriteria } from "../../services/postApi";
+import { Outlet } from "react-router-dom";
 
-function BoardLayout({ category, children }) {
+function BoardLayout({ category = { title: "프로젝트 게시판", id: 3 } }) {
   console.log(category);
   return (
     <Container>
@@ -30,26 +24,7 @@ function BoardLayout({ category, children }) {
       <MainContentWrapper>
         <Navbar />
         <Content>
-          <ContentWrapper>
-            <BoardTitle>{category.title}</BoardTitle>
-            <SearchSortWrapper>
-              <SearchBox>Search</SearchBox>
-              <SortOption
-                onChange={(e) => {
-                  const value = e.target.value;
-                  if (value === "date") {
-                    sortPostsByCriteria(category.id, "date", 0);
-                  } else if (value === "like") {
-                    sortPostsByCriteria(category.id, "like", 0);
-                  }
-                }}
-              >
-                <option value="date">최신순</option>
-                <option value="like">인기순</option>
-              </SortOption>
-            </SearchSortWrapper>
-            <PostCardWrapper>{children}</PostCardWrapper>
-          </ContentWrapper>
+          <Outlet />
 
           <RightSidebarWrapper>
             <PopularCardWrapper>
