+++ src/components/Common/Search/Search.jsx
@@ -1,0 +1,7 @@
+import { SearchBox } from "./Search.style";
+
+function Search() {
+  return <SearchBox>Search</SearchBox>;
+}
+
+export default Search;
