+++ src/components/Common/Search/Search.style.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const SearchBox = styled.div`
+  width: 378px;
+  height: 32px;
+  border: 1px solid blue;
+`;
