+++ src/components/Common/SortOptionButton/SortOptionButton.style.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const SortOption = styled.select`
+  width: 80px;
+  height: 24px;
+  border: 1px solid blue;
+  align-self: flex-end;
+  cursor: pointer;
+`;
